Extract editor create box action
